+++ server/src/main/java/io/whitefox/services/DeltaSharesServiceImpl.java
@@ -55,7 +55,7 @@ public class DeltaSharesServiceImpl implements DeltaSharesService {
   }
 
   @Override
-  public CompletionStage<Optional<Integer>> getTableVersion(
+  public CompletionStage<Optional<Long>> getTableVersion(
       String share, String schema, String table, Optional<String> startingTimestamp) {
     var tableRes = storageManager.getTable(share, schema, table);
     return tableRes.thenApply(t -> t.map(DeltaSharedTable::new)).thenCompose(odt -> odt.map(
